Refactor Form component to follow requirements
